requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.0.33
python-dotenv==1.0.0 